Rename parser to processor
Conflicted with internal Python module, which made imports in test fail
